Updates to transferrable
According to the docs :-

"When people drag and drop a layer within the app or onto another app that recognizes
the custom layer content type, the app uses the first representation. When people drag
and drop the layer onto a different image editor, it’s likely that the editor recognizes the
PNG file type. The second transfer representation adds support for PNG files."

https://developer.apple.com/documentation/coretransferable/transferable